UserLocatableReviews and UserLocatableTips Controllers - SPRINT 4 END

+++ src/main/java/goingto/com/model/interaction/Review.java
@@ -29,7 +29,7 @@ public class Review {
 
     @NotEmpty
     @Column(name = "stars")
-    private Integer stars;
+    private Float stars;
 
     @NotEmpty
     @Column(name = "reviewed_at", length = 10)
